refactor: change storys and results file to match constant name
