+++ src/components/Game/Results.tsx
