+++ src/components/Game/Storys.tsx
